LPS-128860 Move the appender to the proper package
